# Improve router performance with dedicated child types