Before this commit, all the node types were added to the same list of children nodes. Taking in consideration that only one Param and Any type of node could exist per node, two new node struct field were added to hold the references to those kind of nodes.
This avoid the need to iterate through all the Static type nodes just to find one Param or Any type node. Those iterations could be performed multiple times in the same iteration of Router#Find.
Removing the route comments of the Router benchmark tests.
Updating the Router benchmarks tests to find the routes defined to each particular benchmark. Before, all the benchmarks tried to find only the GitHub API.
Adding new router benchmarks to measure when the Router try to find routes that are not registered.

To be fair in the below benchmarks, I did the same modifications to the Router benchmarks on master. Also I performed 10 iterations per each benchmark, and compared the results with benchstat

benchstat bench_master.txt bench_new.txt
```
name                         old time/op    new time/op    delta
RouterStaticRoutes-2           20.0µs ± 2%    20.3µs ± 5%     ~     (p=0.065 n=10+9)
RouterStaticRoutesMisses-2     1.02µs ± 2%    0.66µs ± 4%  -35.43%  (p=0.000 n=8+9)
RouterGitHubAPI-2              38.9µs ± 3%    37.0µs ± 4%   -4.77%  (p=0.000 n=10+10)
RouterGitHubAPIMisses-2        1.25µs ± 3%    0.77µs ± 6%  -38.39%  (p=0.000 n=10+10)
RouterParseAPI-2               2.08µs ±10%    1.98µs ± 3%   -4.71%  (p=0.029 n=10+10)
RouterParseAPIMisses-2          442ns ± 4%     452ns ± 3%     ~     (p=0.066 n=9+9)
RouterGooglePlusAPI-2          1.35µs ± 2%    1.29µs ± 3%   -4.27%  (p=0.000 n=10+10)
RouterGooglePlusAPIMisses-2     989ns ± 4%     662ns ± 8%  -33.06%  (p=0.000 n=10+10)
RouterParamsAndAnyAPI-2        3.34µs ± 1%    2.96µs ± 3%  -11.41%  (p=0.000 n=8+10)

name                         old alloc/op   new alloc/op   delta
RouterStaticRoutes-2            0.00B          0.00B          ~     (all equal)
RouterStaticRoutesMisses-2      0.00B          0.00B          ~     (all equal)
RouterGitHubAPI-2               0.00B          0.00B          ~     (all equal)
RouterGitHubAPIMisses-2         0.00B          0.00B          ~     (all equal)
RouterParseAPI-2                0.00B          0.00B          ~     (all equal)
RouterParseAPIMisses-2          0.00B          0.00B          ~     (all equal)
RouterGooglePlusAPI-2           0.00B          0.00B          ~     (all equal)
RouterGooglePlusAPIMisses-2     0.00B          0.00B          ~     (all equal)
RouterParamsAndAnyAPI-2         0.00B          0.00B          ~     (all equal)

name                         old allocs/op  new allocs/op  delta
RouterStaticRoutes-2             0.00           0.00          ~     (all equal)
RouterStaticRoutesMisses-2       0.00           0.00          ~     (all equal)
RouterGitHubAPI-2                0.00           0.00          ~     (all equal)
RouterGitHubAPIMisses-2          0.00           0.00          ~     (all equal)
RouterParseAPI-2                 0.00           0.00          ~     (all equal)
RouterParseAPIMisses-2           0.00           0.00          ~     (all equal)
RouterGooglePlusAPI-2            0.00           0.00          ~     (all equal)
RouterGooglePlusAPIMisses-2      0.00           0.00          ~     (all equal)
RouterParamsAndAnyAPI-2          0.00           0.00          ~     (all equal)
```