@lammel 2020-12-06
@pafuent I do not see too much risk in changed behaviour as our router tests are pretty helpful here.
So I'm in favour of merging. Patch looks good based on the code, haven't tried the branch yet.

I'm missing our github action for benchmark comparison though.