@pafuent 2020-12-10
Sadly the GitHub action won't be fair in this case, because I changed the benchmarks on this PR. In this case you will need to trust on my benchmarks :wink: